backend/app/core/security.py
"""Security helpers: password hashing and JWT creation/verification."""

from datetime import datetime, timedelta, timezone
from typing import Any, Dict, Optional

from jose import JWTError, jwt
from passlib.context import CryptContext

from .config import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def get_password_hash(password: str) -> str:
    """Return a bcrypt hash for the given plain password."""
    return pwd_context.hash(password)


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a plain password against the stored hash."""
    return pwd_context.verify(plain_password, hashed_password)


def create_access_token(data: Dict[str, Any], expires_delta_minutes: int | None = None) -> str:
    """Create a signed JWT access token."""
    to_encode = data.copy()
    if expires_delta_minutes is None:
        expires_delta_minutes = settings.access_token_expire_minutes
    expire = datetime.now(timezone.utc) + timedelta(minutes=expires_delta_minutes)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)
    return encoded_jwt


def decode_token(token: str) -> Dict[str, Any]:
    """Decode a JWT and return its payload. Raises JWTError on failure."""
    payload: Dict[str, Any] = jwt.decode(token, settings.secret_key, algorithms=[settings.algorithm])
    return payload
