backend/pyproject.toml
[tool.poetry]
name = "primetrade-backend"
version = "0.1.0"
description = "PrimeTrade FastAPI backend"
authors = ["PrimeTrade Team <devnull@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
motor = "^3.3.2"
pymongo = "^4.6.1"
structlog = "^24.1.0"
redis = "^5.0.1"
python-multipart = "^0.0.9"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"
httpx = "^0.27.0"
bcrypt = "<4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
anyio = "^4.2.0"
coverage = "^7.4.0"
ruff = "^0.5.0"

[tool.poetry.scripts]
primetrade-api = "app.main:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["app/tests"]
asyncio_mode = "auto"
